+++ packages/react-native-video/android/src/main/java/com/twg/video/core/utils/VideoOrientationUtils.kt
@@ -3,7 +3,7 @@ package com.twg.video.core.utils
 import com.margelo.nitro.video.VideoOrientation
 
 object VideoOrientationUtils {
-  fun fromWHR(width: Int?, height: Int?, rotation: Int?): VideoOrientation {
+  fun fromWHR(width: Double?, height: Double?, rotation: Int?): VideoOrientation {
     if (width == 0 || height == 0 || height == null || width == null) return VideoOrientation.UNKNOWN
 
     if (width == height) return VideoOrientation.SQUARE
